requirements.txt
aiofiles==0.6.0
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
colorama==0.4.4
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
iniconfig==1.1.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
multidict==5.0.0
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
rfc3986==1.4.0
sanic==20.9.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.22
toml==0.10.2
websockets==8.1
